# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type StoryInputLength = typeof StoryInputLength[keyof typeof StoryInputLength];


export const StoryInputLength = {
  short: 'short',
  long: 'long',
} as const;

export interface StoryInput {
  childName: string;
  age: string;
  storyType: string;
  mood: string;
  language: string;
  characters?: string;
  length?: StoryInputLength;
}

export interface StoryResult {
  story: string;
}

export interface ApiError {
  error: string;
}