# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: stories
    description: Story generation operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /stories/generate:
    post:
      operationId: generateStory
      tags: [stories]
      summary: Generate an AI story for a child
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoryInput"
      responses:
        "200":
          description: Generated story
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoryResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    StoryInput:
      type: object
      properties:
        childName:
          type: string
        age:
          type: string
        storyType:
          type: string
        mood:
          type: string
        language:
          type: string
        characters:
          type: string
        length:
          type: string
          enum: [short, long]
      required:
        - childName
        - age
        - storyType
        - mood
        - language
    StoryResult:
      type: object
      properties:
        story:
          type: string
      required:
        - story
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error