# lib/api-zod/src/generated/types/storyInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StoryInputLength } from './storyInputLength';

export interface StoryInput {
  childName: string;
  age: string;
  storyType: string;
  mood: string;
  language: string;
  characters?: string;
  length?: StoryInputLength;
}